finish functions refactoring
