+++ dbx/commands/sync/sync.py
@@ -1,14 +1,12 @@
-import os
 from typing import List
 
 import click
-from databricks_cli.configure.provider import DatabricksConfig
 
-from dbx.commands.sync.functions import validate_allow_unmatched, create_path_matcher, main_loop
+from dbx.commands.sync.functions import validate_allow_unmatched, create_path_matcher, main_loop, handle_source, \
+    get_user_name, get_source_base_name
 from dbx.sync import DeleteUnmatchedOption
-from dbx.sync.clients import DBFSClient, ReposClient, get_user
+from dbx.sync.clients import DBFSClient, ReposClient
 from dbx.sync.config import get_databricks_config
-from dbx.utils import dbx_echo
 
 
 @click.group(short_help=":arrows_counterclockwise: Sync local files to Databricks and watch for changes")
@@ -77,55 +75,6 @@ def sync():
 
     """
     pass  # pragma: no cover
-
-
-def handle_source(source: str = None) -> str:
-    """Determine the source directory to sync from.  If the source directory is not specified
-    then it will check if the current directory has a .git subdirectory, implying that the current
-    directory should be the source.
-
-    Args:
-        source (str): source directory to use, or None to try automatically determining
-
-    Raises:
-        click.UsageError: When the source directory is unspecified and can't be automatically determined.
-
-    Returns:
-        str: source directory to sync from
-    """
-    if not source:
-        # If this is being run from the base of a git repo, then they probably want to use
-        # the repo as the source.
-        if os.path.exists(".git"):
-            source = "."
-        else:
-            raise click.UsageError("Must specify source directory using --source")
-
-    source = os.path.abspath(source)
-
-    dbx_echo(f"Syncing from {source}")
-
-    return source
-
-
-def get_user_name(config: DatabricksConfig) -> str:
-    """Gets the name of the user according to the Databricks API using the config for authorization.
-
-    Args:
-        config (DatabricksConfig): config to use to get user info
-
-    Returns:
-        str: name of user
-    """
-    user_info = get_user(config)
-    return user_info.get("userName")
-
-
-def get_source_base_name(source: str) -> str:
-    source_base_name = os.path.basename(source.rstrip("/"))
-    if not source_base_name:
-        raise click.UsageError("Destination path can't be determined.  Please specify with --dest.")
-    return source_base_name
 
 
 def common_options(f):
